+++ src/wetterdienst/provider/aemet/observation/__init__.py
@@ -1,5 +1,6 @@
 # Copyright (C) 2018-2025, earthobservations developers.
 # Distributed under the MIT License. See LICENSE for more info.
 from wetterdienst.provider.aemet.observation.api import AemetObservationRequest
+from wetterdienst.provider.aemet.observation.metadata import AemetObservationMetadata
 
-__all__ = ["AemetObservationRequest"]
+__all__ = ["AemetObservationMetadata", "AemetObservationRequest"]