+++ src/wetterdienst/provider/smhi/observation/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (C) 2018-2025, earthobservations developers.
+# Distributed under the MIT License. See LICENSE for more info.
+from wetterdienst.provider.smhi.observation.api import SmhiObservationRequest
+from wetterdienst.provider.smhi.observation.metadata import SmhiObservationMetadata
+
+__all__ = ["SmhiObservationMetadata", "SmhiObservationRequest"]