+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Types of changes:
   annual resolution (no authentication required)
 - Add AEMET (Spain) observation provider with hourly (real-time), daily, monthly and
   annual resolution
+- Add SMHI (Sweden) observation provider with 1-minute, hourly, daily and monthly resolution
 - Add Météo-France (France) synop network (subdaily, 3-hourly)
 - Add Météo-France (France) observation network (6-minute, hourly, daily, monthly)
 - Add MeteoSwiss (Switzerland) observation provider with 10-minute, hourly, daily, monthly and annual resolution